requirements.txt
streamlit
google-generativeai>=0.4.0
langchain
langchain-core
langchain-community
langchain-google-genai
faiss-cpu
pypdf
sentence-transformers
